fix(filters): resolved ReDoS vulnerability
