+++ src/.env
@@ -57,8 +57,6 @@ WEBHOOK_ERROR_ID='648663047615021058'
 WEBHOOK_ERROR_NAME='Skyra'
 WEBHOOK_ERROR_TOKEN=''
 
-WORKER_COUNT=2
-
 # Tokens
 DISCORD_TOKEN=''
 BOTLIST_SPACE_TOKEN=''
