+++ src/events/moderation/messages/guildUserMessageWordsHandler.ts
@@ -2,12 +2,12 @@ import { GuildSettings } from '#lib/database';
 import { SkyraEmbed } from '#lib/discord';
 import { LanguageKeys } from '#lib/i18n/languageKeys';
 import { ModerationMessageEvent } from '#lib/moderation';
-import { IncomingType, OutgoingType } from '#lib/moderation/workers';
 import type { GuildMessage } from '#lib/types';
 import { Colors } from '#lib/types/Constants';
 import { floatPromise, getContent } from '#utils/util';
 import { ApplyOptions } from '@sapphire/decorators';
 import { codeBlock, cutText } from '@sapphire/utilities';
+import { remove as removeConfusables } from 'confusables';
 import type { TextChannel } from 'discord.js';
 import type { TFunction } from 'i18next';
 
@@ -29,11 +29,8 @@ export class UserModerationMessageEvent extends ModerationMessageEvent {
 		const content = getContent(message);
 		if (content === null) return null;
 
-		const regExp = await message.guild.readSettings((settings) => settings.wordFilterRegExp);
-		if (regExp === null) return null;
-
-		const result = await this.context.workers.send({ type: IncomingType.RunRegExp, regExp, content });
-		return result.type === OutgoingType.RegExpMatch ? result : null;
+		const regex = await message.guild.readSettings((settings) => settings.wordFilterRegExp);
+		return regex ? this.filter(removeConfusables(content), regex) : null;
 	}
 
 	protected async onDelete(message: GuildMessage, t: TFunction, value: FilterResults) {
@@ -56,6 +53,40 @@ export class UserModerationMessageEvent extends ModerationMessageEvent {
 			.setAuthor(`${message.author.tag} (${message.author.id})`, message.author.displayAvatarURL({ size: 128, format: 'png', dynamic: true }))
 			.setFooter(`#${(message.channel as TextChannel).name} | ${t(LanguageKeys.Events.Moderation.Messages.WordFooter)}`)
 			.setTimestamp();
+	}
+
+	private filter(str: string, regex: RegExp): FilterResults | null {
+		const matches = str.match(regex);
+		if (matches === null) return null;
+
+		let last = 0;
+		let next = 0;
+
+		const filtered: string[] = [];
+		const highlighted: string[] = [];
+		for (const match of matches) {
+			next = str.indexOf(match, last);
+			const section = str.slice(last, next);
+			if (section) {
+				filtered.push(section, '*'.repeat(match.length));
+				highlighted.push(section, `__${match}__`);
+			} else {
+				filtered.push('*'.repeat(match.length));
+				highlighted.push(`__${match}__`);
+			}
+			last = next + match.length;
+		}
+
+		if (last !== str.length) {
+			const end = str.slice(last);
+			filtered.push(end);
+			highlighted.push(end);
+		}
+
+		return {
+			filtered: filtered.join(''),
+			highlighted: highlighted.join('')
+		};
 	}
 }
