--- src/lib/moderation/workers/errors/TimeoutError.ts
@@ -1,5 +1,0 @@
-export class TimeoutError extends Error {
-	public constructor() {
-		super('Reached Timeout');
-	}
-}
