--- src/lib/moderation/workers/errors/index.ts
@@ -1,1 +1,0 @@
-export * from './TimeoutError';
