--- src/lib/moderation/workers/index.ts
@@ -1,5 +1,0 @@
-export * from './errors';
-export * from './types';
-export * from './worker';
-export * from './WorkerHandler';
-export * from './WorkerResponseHandler';
