--- src/lib/moderation/workers/types.ts
@@ -1,39 +1,0 @@
-export interface IdentifiablePayload {
-	id: number;
-}
-
-export type NoId<T> = Omit<T, 'id'>;
-
-export const enum IncomingType {
-	RunRegExp
-}
-
-export type IncomingPayload = IncomingRunRegExpPayload;
-
-export interface IncomingRunRegExpPayload extends IdentifiablePayload {
-	type: IncomingType.RunRegExp;
-	regExp: RegExp;
-	content: string;
-}
-
-export const enum OutgoingType {
-	Heartbeat,
-	NoContent,
-	RegExpMatch
-}
-
-export type OutgoingPayload = OutgoingHeartbeatPayload | OutgoingNoContentPayload | OutgoingRegExpMatchPayload;
-
-export interface OutgoingHeartbeatPayload {
-	type: OutgoingType.Heartbeat;
-}
-
-export interface OutgoingNoContentPayload extends IdentifiablePayload {
-	type: OutgoingType.NoContent;
-}
-
-export interface OutgoingRegExpMatchPayload extends IdentifiablePayload {
-	type: OutgoingType.RegExpMatch;
-	filtered: string;
-	highlighted: string;
-}
