+++ src/lib/types/Augments.d.ts
@@ -2,7 +2,6 @@
 import type { NP, Queue, QueueClient, QueueClientOptions, QueueEntry } from '#lib/audio';
 import type { DbSet, SettingsManager } from '#lib/database';
 import type { GuildMemberFetchQueue } from '#lib/discord/GuildMemberFetchQueue';
-import type { WorkerManager } from '#lib/moderation/workers/WorkerManager';
 import type { AnalyticsData, ColorHandler, GiveawayManager, InviteCodeValidEntry, InviteStore, ScheduleManager, SkyraCommand } from '#lib/structures';
 import type { AnalyticsSchema } from '#lib/types/AnalyticsSchema';
 import type { WebsocketHandler } from '#root/audio/lib/websocket/WebsocketHandler';
@@ -117,7 +116,6 @@ declare module '@sapphire/pieces' {
 	interface PieceContextExtras {
 		db: DbSet;
 		schedule: ScheduleManager;
-		workers: WorkerManager;
 	}
 }
